refactor: extract deletePart function from useGetPart to a dedicated useDeletePart hook
